requirements.txt
python-telegram-bot==22.3
openai
nest_asyncio
python-dotenv
